Optimize d14-1.c to reduce the amount of passes made
It was blindly incrementing the fuel needed by 1, until it hit
1 trillion ORE. But that was taking almost 2.6 million loops, and
the run_list can be costly.  So, we run it once and get an amount
to make 1 FUEL. We use that as our initial estimate for how much
we can make (1 Trillion / cost_of_1_FUEL). This will always be
lower than the real amount we can make due to efficiencies in the
extra materials we make along the way.  We then make that many
fuel.  We use that new amount and its cost to find the average
ORE needed for 1 FUEL. This number is so close to the real amount
that we need to increment it because we lose the remainder and
will go over if we use it.

Then, we use the new average cost to estimate how many FUEL we
can make. This gets us almost exactly there, and we can loop by
1 until we go over. We actually are 1 under the real amount at
this point, so 2 loops gets us over and the program runs almost
instantly now.